+++ movie_list/src/index.css
@@ -1,30 +1,13 @@
-
 body {
   margin: 0;
-  min-height: 100vh;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  display: flex;
-  background-color: #FCE694;
-
 }
 
-#root {
-  flex-grow: 1;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-
-.App {
-  width: 100%;
-  max-width: 800px;
-  height: 100vh;
-  display: flex;
-  justify-content: flex-start;
-  align-items: center;
+code {
+  font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
+    monospace;
 }
