+++ package.json
@@ -4,3 +4,4 @@
     "server": "^1.0.37"
   }
 }
+inst
